refactor ReferendumDisplayStringFactory
