+++ novawallet/Modules/Vote/Governance/Delegations/DelegateInfo/GovernanceDelegateInfoViewFactory.swift
@@ -18,9 +18,7 @@ struct GovernanceDelegateInfoViewFactory {
 
         let localizationManager = LocalizationManager.shared
 
-        let referendumDisplayStringFactory = ReferendumDisplayStringFactoryProvider.factory(
-            for: state.settings.value.type
-        )
+        let referendumDisplayStringFactory = ReferendumDisplayStringFactory()
 
         let governanceDelegateInfoViewModelFactory = GovernanceDelegateInfoViewModelFactory(
             stringDisplayFactory: referendumDisplayStringFactory
