+++ novawallet/Modules/Vote/Governance/Delegations/DelegateSearch/GovernanceDelegateSearchViewFactory.swift
@@ -14,9 +14,7 @@ struct GovernanceDelegateSearchViewFactory {
             return nil
         }
 
-        let referendumDisplayStringFactory = ReferendumDisplayStringFactoryProvider.factory(
-            for: state.settings.value.type
-        )
+        let referendumDisplayStringFactory = ReferendumDisplayStringFactory()
         let addressViewModelFactory = DisplayAddressViewModelFactory()
         let quantityFormatter = NumberFormatter.quantity.localizableResource()
