+++ novawallet/Modules/Vote/Governance/ReferendumDetails/ReferendumDetailsPresenter.swift
@@ -9,7 +9,7 @@ final class ReferendumDetailsPresenter {
     let balanceViewModelFactory: BalanceViewModelFactoryProtocol
     let referendumFormatter: LocalizableResource<NumberFormatter>
     let referendumViewModelFactory: ReferendumsModelFactoryProtocol
-    let referendumStringsFactory: ReferendumDisplayStringFactoryProtocol
+    let referendumVotesFactory: ReferendumVotesViewModelFactoryProtocol
     let referendumTimelineViewModelFactory: ReferendumTimelineViewModelFactoryProtocol
     let referendumMetadataViewModelFactory: ReferendumMetadataViewModelFactoryProtocol
     let endedReferendumProgressViewModelFactory: EndedReferendumProgressViewModelFactoryProtocol?
@@ -56,7 +56,7 @@ final class ReferendumDetailsPresenter {
         referendumViewModelFactory: ReferendumsModelFactoryProtocol,
         balanceViewModelFactory: BalanceViewModelFactoryProtocol,
         referendumFormatter: LocalizableResource<NumberFormatter>,
-        referendumStringsFactory: ReferendumDisplayStringFactoryProtocol,
+        referendumVotesFactory: ReferendumVotesViewModelFactoryProtocol,
         referendumTimelineViewModelFactory: ReferendumTimelineViewModelFactoryProtocol,
         referendumMetadataViewModelFactory: ReferendumMetadataViewModelFactoryProtocol,
         endedReferendumProgressViewModelFactory: EndedReferendumProgressViewModelFactoryProtocol?,
@@ -71,7 +71,7 @@ final class ReferendumDetailsPresenter {
         self.governanceType = governanceType
         self.accountManagementFilter = accountManagementFilter
         self.referendumViewModelFactory = referendumViewModelFactory
-        self.referendumStringsFactory = referendumStringsFactory
+        self.referendumVotesFactory = referendumVotesFactory
         self.balanceViewModelFactory = balanceViewModelFactory
         self.referendumFormatter = referendumFormatter
         self.referendumTimelineViewModelFactory = referendumTimelineViewModelFactory
@@ -248,7 +248,7 @@ extension ReferendumDetailsPresenter {
         let viewModel: [YourVoteRow.Model]
 
         if let accountVotes = accountVotes {
-            viewModel = referendumStringsFactory.createDirectVotesViewModel(
+            viewModel = referendumVotesFactory.createDirectVotesViewModel(
                 from: accountVotes,
                 chain: chain,
                 locale: selectedLocale
@@ -256,13 +256,13 @@ extension ReferendumDetailsPresenter {
         } else if let offchainVoting = offchainVoting {
             switch offchainVoting.voteType {
             case let .direct(directVote):
-                viewModel = referendumStringsFactory.createDirectVotesViewModel(
+                viewModel = referendumVotesFactory.createDirectVotesViewModel(
                     from: directVote,
                     chain: chain,
                     locale: selectedLocale
                 )
             case let .delegated(delegateVote):
-                viewModel = referendumStringsFactory.createDelegatorVotesViaDelegateViewModel(
+                viewModel = referendumVotesFactory.createDelegatorVotesViaDelegateViewModel(
                     from: delegateVote,
                     delegateName: offchainVoting.identity?.displayName ?? offchainVoting.metadata?.name,
                     chain: chain,
@@ -339,7 +339,7 @@ extension ReferendumDetailsPresenter {
             )
         )
 
-        let votes = referendumStringsFactory.createReferendumVotes(
+        let votes = referendumVotesFactory.createReferendumVotes(
             from: referendum,
             offchainVotingAmount: offchainVotingAmount,
             chain: chain,
