+++ novawallet/Modules/Vote/Governance/ViewModel/ReferendumDisplayStringFactory.swift
@@ -2,9 +2,19 @@ import Foundation
 import BigInt
 
 protocol ReferendumDisplayStringFactoryProtocol {
-    func createVotesValue(from votes: BigUInt, chain: ChainModel, locale: Locale) -> String?
+    var formatterFactory: AssetBalanceFormatterFactoryProtocol { get }
 
-    func createVotes(from votes: BigUInt, chain: ChainModel, locale: Locale) -> String?
+    func createVotesValue(
+        from votes: BigUInt,
+        chain: ChainModel,
+        locale: Locale
+    ) -> String?
+
+    func createVotes(
+        from votes: BigUInt,
+        chain: ChainModel,
+        locale: Locale
+    ) -> String?
 
     func createVotesDetails(
         from amount: BigUInt,
@@ -12,310 +22,9 @@ protocol ReferendumDisplayStringFactoryProtocol {
         chain: ChainModel,
         locale: Locale
     ) -> String?
-
-    func createReferendumVotes(
-        from referendum: ReferendumLocal,
-        offchainVotingAmount: ReferendumVotingAmount?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> ReferendumVotesViewModel
-
-    func createDirectVotesViewModel(
-        from vote: ReferendumAccountVoteLocal,
-        chain: ChainModel,
-        locale: Locale
-    ) -> [YourVoteRow.Model]
-
-    func createDelegatorVotesViaDelegateViewModel(
-        from vote: GovernanceOffchainVoting.DelegateVote,
-        delegateName: String?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> [YourVoteRow.Model]
 }
 
-private typealias ReferendumDisplayStringFactory = ReferendumDisplayStringFactoryProtocol & BaseReferendumDisplayStringFactory
-
-enum ReferendumDisplayStringFactoryProvider {
-    static func factory(
-        for govType: GovernanceType,
-        formatterFactory: AssetBalanceFormatterFactoryProtocol = AssetBalanceFormatterFactory()
-    ) -> ReferendumDisplayStringFactoryProtocol {
-        if govType == .governanceV2 {
-            OpenGovReferendumDisplayStringFactory(formatterFactory: formatterFactory)
-        } else {
-            Gov1ReferendumDisplayStringFactory(formatterFactory: formatterFactory)
-        }
-    }
-}
-
-// MARK: OpenGov Factory
-
-final class OpenGovReferendumDisplayStringFactory: ReferendumDisplayStringFactory {
-    let formatterFactory: AssetBalanceFormatterFactoryProtocol
-
-    init(formatterFactory: AssetBalanceFormatterFactoryProtocol = AssetBalanceFormatterFactory()) {
-        self.formatterFactory = formatterFactory
-    }
-
-    func createViewModel(
-        title: String,
-        value: BigUInt?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> VoteRowView.Model? {
-        let loadableValueString: LoadableViewModelState<String>
-
-        if
-            let value,
-            let valueString = createVotes(
-                from: value,
-                chain: chain,
-                locale: locale
-            ) {
-            loadableValueString = .loaded(value: valueString)
-        } else {
-            loadableValueString = .loading
-        }
-
-        let viewModel = VoteRowView.Model(
-            title: title,
-            votes: loadableValueString
-        )
-
-        return viewModel
-    }
-}
-
-// MARK: Gov1 Factory
-
-final class Gov1ReferendumDisplayStringFactory: ReferendumDisplayStringFactory {
-    let formatterFactory: AssetBalanceFormatterFactoryProtocol
-
-    init(formatterFactory: AssetBalanceFormatterFactoryProtocol = AssetBalanceFormatterFactory()) {
-        self.formatterFactory = formatterFactory
-    }
-
-    func createViewModel(
-        title: String,
-        value: BigUInt?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> VoteRowView.Model? {
-        guard
-            let value,
-            let valueString = createVotes(
-                from: value,
-                chain: chain,
-                locale: locale
-            )
-        else {
-            return nil
-        }
-
-        let loadableValueString: LoadableViewModelState<String> = .loaded(value: valueString)
-
-        let viewModel = VoteRowView.Model(
-            title: title,
-            votes: loadableValueString
-        )
-
-        return viewModel
-    }
-}
-
-// MARK: BaseReferendumDisplayStringFactory
-
-protocol BaseReferendumDisplayStringFactory {
-    var formatterFactory: AssetBalanceFormatterFactoryProtocol { get }
-
-    func createViewModel(
-        title: String,
-        value: BigUInt?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> VoteRowView.Model?
-}
-
-extension BaseReferendumDisplayStringFactory {
-    func createReferendumVotes(
-        from referendum: ReferendumLocal,
-        offchainVotingAmount: ReferendumVotingAmount?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> ReferendumVotesViewModel {
-        var ayes: BigUInt?
-        var nays: BigUInt?
-        var abstains: BigUInt? = offchainVotingAmount?.abstain
-
-        if let voting = referendum.voting {
-            switch voting {
-            case let .supportAndVotes(model):
-                ayes = model.ayes
-                nays = model.nays
-            case let .threshold(model):
-                ayes = model.ayes
-                nays = model.nays
-            }
-        } else if let offchainVotingAmount {
-            ayes = offchainVotingAmount.aye
-            nays = offchainVotingAmount.nay
-        }
-
-        let aye = createViewModel(
-            title: R.string.localizable.governanceAye(preferredLanguages: locale.rLanguages),
-            value: ayes,
-            chain: chain,
-            locale: locale
-        )
-        let nay = createViewModel(
-            title: R.string.localizable.governanceNay(preferredLanguages: locale.rLanguages),
-            value: nays,
-            chain: chain,
-            locale: locale
-        )
-        var abstain = createViewModel(
-            title: R.string.localizable.governanceAbstain(preferredLanguages: locale.rLanguages),
-            value: abstains,
-            chain: chain,
-            locale: locale
-        )
-
-        return ReferendumVotesViewModel(
-            ayes: aye,
-            nays: nay,
-            abstains: abstain
-        )
-    }
-
-    func createVotesValue(from votes: BigUInt, chain: ChainModel, locale: Locale) -> String? {
-        guard let asset = chain.utilityAsset() else {
-            return nil
-        }
-
-        let displayInfo = ChainAsset(chain: chain, asset: asset).assetDisplayInfo
-
-        let votesDecimal = Decimal.fromSubstrateAmount(votes, precision: displayInfo.assetPrecision) ?? 0
-
-        let displayFormatter = formatterFactory.createDisplayFormatter(for: displayInfo).value(for: locale)
-
-        return displayFormatter.stringFromDecimal(votesDecimal)
-    }
-
-    func createDirectVotesViewModel(
-        from vote: ReferendumAccountVoteLocal,
-        chain: ChainModel,
-        locale: Locale
-    ) -> [YourVoteRow.Model] {
-        var viewModels: [YourVoteRow.Model] = []
-
-        if vote.hasAyeVotes {
-            let viewModel = createYourVoteRowViewModel(
-                vote: .init(balance: vote.ayeBalance, conviction: vote.convictionValue),
-                typeName: R.string.localizable.governanceAye(preferredLanguages: locale.rLanguages),
-                style: .ayeInverse,
-                chain: chain,
-                locale: locale
-            )
-
-            viewModels.append(viewModel)
-        }
-
-        if vote.hasNayVotes {
-            let viewModel = createYourVoteRowViewModel(
-                vote: .init(balance: vote.nayBalance, conviction: vote.convictionValue),
-                typeName: R.string.localizable.governanceNay(preferredLanguages: locale.rLanguages),
-                style: .nayInverse,
-                chain: chain,
-                locale: locale
-            )
-
-            viewModels.append(viewModel)
-        }
-
-        if vote.hasAbstainVotes {
-            let viewModel = createYourVoteRowViewModel(
-                vote: .init(balance: vote.abstainBalance, conviction: vote.convictionValue),
-                typeName: R.string.localizable.governanceAbstain(preferredLanguages: locale.rLanguages),
-                style: .abstainInverse,
-                chain: chain,
-                locale: locale
-            )
-
-            viewModels.append(viewModel)
-        }
-
-        return viewModels
-    }
-
-    func createDelegatorVotesViaDelegateViewModel(
-        from vote: GovernanceOffchainVoting.DelegateVote,
-        delegateName: String?,
-        chain: ChainModel,
-        locale: Locale
-    ) -> [YourVoteRow.Model] {
-        let votesValue = vote.delegatorPower.conviction.votes(for: vote.delegatorPower.balance) ?? 0
-
-        let votesString = createVotes(
-            from: votesValue,
-            chain: chain,
-            locale: locale
-        )
-
-        let voteSideString: String
-        let voteSideStyle: YourVoteView.Style
-
-        if vote.delegateVote.vote.aye {
-            voteSideString = R.string.localizable.governanceAye(preferredLanguages: locale.rLanguages)
-            voteSideStyle = .ayeInverse
-        } else {
-            voteSideString = R.string.localizable.governanceNay(preferredLanguages: locale.rLanguages)
-            voteSideStyle = .nayInverse
-        }
-
-        let delegateName = delegateName ?? vote.delegateAddress
-
-        let voteDescription = R.string.localizable.delegatorVotesViaDelegate(
-            delegateName,
-            preferredLanguages: locale.rLanguages
-        )
-
-        let viewModel = YourVoteRow.Model(
-            vote: .init(title: voteSideString.uppercased(), description: voteDescription, style: voteSideStyle),
-            amount: .init(topValue: votesString ?? "", bottomValue: nil)
-        )
-
-        return [viewModel]
-    }
-
-    func createYourVoteRowViewModel(
-        vote: GovernanceBalanceConviction,
-        typeName: String,
-        style: YourVoteView.Style,
-        chain: ChainModel,
-        locale: Locale
-    ) -> YourVoteRow.Model {
-        let votesString = createVotes(
-            from: vote.votes ?? 0,
-            chain: chain,
-            locale: locale
-        )
-
-        let convictionString = createVotesDetails(
-            from: vote.balance,
-            conviction: vote.conviction.decimalValue,
-            chain: chain,
-            locale: locale
-        )
-
-        let voteDescription = R.string.localizable.govYourVote(preferredLanguages: locale.rLanguages)
-
-        return YourVoteRow.Model(
-            vote: .init(title: typeName.uppercased(), description: voteDescription, style: style),
-            amount: .init(topValue: votesString ?? "", bottomValue: convictionString)
-        )
-    }
-
+extension ReferendumDisplayStringFactoryProtocol {
     func createVotes(
         from votes: BigUInt,
         chain: ChainModel,
@@ -361,5 +70,27 @@ extension BaseReferendumDisplayStringFactory {
         } else {
             return nil
         }
+    }
+}
+
+struct ReferendumDisplayStringFactory: ReferendumDisplayStringFactoryProtocol {
+    let formatterFactory: AssetBalanceFormatterFactoryProtocol
+
+    init(formatterFactory: AssetBalanceFormatterFactoryProtocol = AssetBalanceFormatterFactory()) {
+        self.formatterFactory = formatterFactory
+    }
+
+    func createVotesValue(from votes: BigUInt, chain: ChainModel, locale: Locale) -> String? {
+        guard let asset = chain.utilityAsset() else {
+            return nil
+        }
+
+        let displayInfo = ChainAsset(chain: chain, asset: asset).assetDisplayInfo
+
+        let votesDecimal = Decimal.fromSubstrateAmount(votes, precision: displayInfo.assetPrecision) ?? 0
+
+        let displayFormatter = formatterFactory.createDisplayFormatter(for: displayInfo).value(for: locale)
+
+        return displayFormatter.stringFromDecimal(votesDecimal)
     }
 }
